Add two components and implement front-end routing
Add the About and Terms components and use React BrowserRouter, Switch
and Route to implement cliend-side/front-end routing.

This way we can avoid the need to download and render entire pages,
because JavaScript can load content in the background and display it
when it is ready, updating only the parts of the page that need to be
updated.

In addition, we can give the illusion of traditional navigation links,
allowing the user to bookmark pages and use the browser's back and
forward buttons.