Handle the submit form
Use React useState to get the values from the form. useState() returns a
state value and a function to update it.

const [state, setState] = useState().

Inside the form, the onChange event calls the setUsername, passing as
its argument the user typed value. The same is valid to the email and
password fields.

Then, with Axios, those values are sent to the server.