Display followers/following through a single component
Use a single component to display the number of followers/following of
a user. This was done since the ProfileFollowers and ProfileFollowing
files had the same content, except for the URL of the Axios request.

The downside of this approach is that everytime the user switches
between tabs, the componet must send a request to the server in order to
display the appropriate content.