Apply lazy loading to the Search and Chat components
Note: CSSTransition will apply classes to its direct children. If you
use it with lazy loading, it will try to apply classes to a components
that was not mounted. A work around it is to simple add a div that will
receive the classes and inside that div put the lazy loaded component.